+++ Real-Time-Animation_3/src/game_objects/Bone.h
@@ -1,12 +1,12 @@
 #pragma once
-#include "core\Application.h"
+#include "Engine.h"
 
 class Bone : public GameObject {
 public:
 	Bone(std::string name, Application * app, Model* model, glm::vec3 object_color, float length);
 	~Bone();
 
-	void update(float delta_time) override;
+	void update(double delta_time) override;
 	void render() override;
 	void set_shader_program(ShaderProgram * shader_program) override;
 	void set_parent(GameObject* parent) override;
